tutorial results description and other fixes for PR
